+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export { parseTtl, DEFAULT_TTL } from "./ttl";
+export { logRequest, logServerRequest, type LogEntry } from "./logger";